Can now generate 2d histograms :)

+++ src/analysis/my_analysis.py
@@ -1,13 +1,14 @@
 import pandas as pd
-from typing import Optional, List, Union
+from typing import Optional, List, Union, Tuple
 import os
 from numpy import pi
+from itertools import combinations
 
 # project imports
 from utils.cutflow import Cutflow
 from utils.cutfile_utils import parse_cutfile, gen_cutgroups, compare_cutfile_backup, backup_cutfile
 from utils.file_utils import identical_to_backup, delete_file, is_dir_empty, get_last_backup, get_filename, makedir
-from utils.plotting_utils import plot_overlay_and_acceptance_cutgroups
+from utils.plotting_utils import plot_1d_overlay_and_acceptance_cutgroups, plot_2d_cutgroups
 from utils.dataframe_utils import (build_analysis_dataframe, create_cut_columns,
                                    gen_weight_column, rescale_to_gev,
                                    get_cross_section, get_luminosity)
@@ -18,7 +19,7 @@ class Analysis:
     # ========= SETUP ===========
     # ===========================
     # multithreading
-    n_threads = os.cpu_count() // 2
+    _n_threads = os.cpu_count() // 2
 
     # options
     TTree = 'truth'  # name of TTree to extract from root file
@@ -42,8 +43,12 @@ class Analysis:
         # set
         self._lepton_name = lepton
         self._cutfile = cutfile
+        self._not_log = [
+            '_phi_',
+            '_eta_',
+        ]
 
-        # special binning
+        # variables that require special (default) binning
         self._special_binning = {
             '_eta_': etabins,
             '_phi_': phibins,
@@ -104,14 +109,17 @@ class Analysis:
         # ===============================
         # ========== CUTFLOW ============
         # ===============================
-        self.Cutflow = Cutflow(self.tree_df, self.cut_dicts, self.cut_label)
+        # don't product cutflow (for now) if not sequential
+        # TODO: non-sequential cutflow
+        if self.options['sequential']:
+            self.Cutflow = Cutflow(self.tree_df, self.cut_dicts, self.cut_label)
 
-        # plot latex table if it doesn't exist and is different to the last file
-        if self._make_backup or is_dir_empty(self.latex_table_dir):
-            last_backup = get_last_backup(self.latex_table_dir)
-            latex_file = self.Cutflow.print_latex_table(self.latex_table_dir)
-            if identical_to_backup(latex_file, backup_file=last_backup):
-                delete_file(latex_file)
+            # plot latex table if it doesn't exist and is different to the last file
+            if self._make_backup or is_dir_empty(self.latex_table_dir):
+                last_backup = get_last_backup(self.latex_table_dir)
+                latex_file = self.Cutflow.print_latex_table(self.latex_table_dir, cutfile_name)
+                if identical_to_backup(latex_file, backup_file=last_backup):
+                    delete_file(latex_file)
 
         # if new cutfile, save backup
         if self._make_backup:
@@ -121,9 +129,11 @@ class Analysis:
     # =========== PLOTS =============
     # ===============================
     # TODO: save outputs in separate directory for each new cutfile
-    def plot_with_cuts(self, scaling: Optional[str] = None,
+    def plot_with_cuts(self,
+                       scaling: Optional[str] = None,
                        bins: Union[tuple, list] = (30, 1, 500),
-                       not_log_add: Optional[List[str]] = None) -> None:
+                       not_log_add: Optional[List[str]] = None
+                       ) -> None:
         """
         Plots each variable to cut from _cutfile with each cutgroup applied
 
@@ -137,39 +147,29 @@ class Analysis:
         """
         # any of the substrings in this list shouldn't be binned logarithmically
         # (may need to double check this as it can cause problems if the substrings appear elsewhere)
-        not_log = [
-            '_phi_',
-            '_eta_',
-        ]
         if not_log_add:
-            not_log += not_log_add
+            self._not_log += not_log_add
 
         for var_to_plot in self.vars_to_cut:
             # whether or not bins should be logarithmic bins
-            is_logbins = not any(map(var_to_plot.__contains__, not_log))
-            if not is_logbins:
-                # set bins for special variables
-                sp_var = [sp_var for sp_var in not_log if sp_var in var_to_plot]
-                if len(sp_var) != 1:
-                    raise Exception(f"Expected one matching variable for spcial binning. Got {sp_var}")
-                in_bins = self._special_binning[sp_var[0]]
-            else:
+            is_logbins, in_bins = self.__getbins(var_to_plot)
+            if not in_bins:
                 in_bins = bins
 
             print(f"Generating histogram for {var_to_plot}...")
-            plot_overlay_and_acceptance_cutgroups(df=self.tree_df,
-                                                  var_to_plot=var_to_plot,
-                                                  weights=self.tree_df['weight'],
-                                                  cutgroups=self.cutgroups,
-                                                  lumi=self.luminosity,
-                                                  dir_path=self.plot_dir,
-                                                  cut_label=self.cut_label,
-                                                  is_logbins=is_logbins,
-                                                  n_threads=self.n_threads,
-                                                  lepton=self._lepton_name,
-                                                  scaling=scaling,
-                                                  bins=in_bins,
-                                                  )
+            plot_1d_overlay_and_acceptance_cutgroups(
+                df=self.tree_df,
+                var_to_plot=var_to_plot,
+                cutgroups=self.cutgroups,
+                lumi=self.luminosity,
+                dir_path=self.plot_dir,
+                cut_label=self.cut_label,
+                is_logbins=is_logbins,
+                n_threads=self._n_threads,
+                lepton=self._lepton_name,
+                scaling=scaling,
+                bins=in_bins,
+            )
 
     def gen_cutflow_hist(self,
                          event: bool = True,
@@ -178,6 +178,9 @@ class Analysis:
                          a_ratio: bool = False
                          ):
         """Generates and saves cutflow histograms"""
+        if not self.options['sequential']:
+            raise Exception("No cutflow available for non-sequential cuts")
+
         if event:
             self.Cutflow.print_histogram(self.plot_dir, 'event')
         if ratio:
@@ -187,13 +190,36 @@ class Analysis:
         if a_ratio:
             self.Cutflow.print_histogram(self.plot_dir, 'a_ratio')
 
-    # def make_all_2dplots(self):
+    def make_all_cutgroup_2dplots(self, bins: Union[tuple, list] = (20, 0, 200)):
+        if len(self.vars_to_cut) < 2:
+            raise Exception("Need at least two plotting variables to make 2D plot")
+
+        for x_var, y_var in combinations(self.vars_to_cut, 2):
+            # binning
+            _, xbins = self.__getbins(x_var)
+            _, ybins = self.__getbins(y_var)
+            if not xbins:
+                xbins = bins
+            if not ybins:
+                ybins = bins
+            print(f"Generating 2d histogram for {x_var}-{y_var}...")
+            plot_2d_cutgroups(self.tree_df,
+                              x_var=x_var, y_var=y_var,
+                              xbins=xbins, ybins=ybins,
+                              cutgroups=self.cutgroups,
+                              dir_path=self.plot_dir,
+                              cut_label=self.cut_label,
+                              lepton=self._lepton_name,
+                              n_threads=self._n_threads,
+                              )
 
     # ===============================
     # ========= PRINTOUTS ===========
     # ===============================
     def cutflow_printout(self) -> None:
         """Prints cutflow table to terminal"""
+        if not self.options['sequential']:
+            raise Exception("No cutflow available for non-sequential cuts")
         self.Cutflow.terminal_printout()
 
     def kinematics_printouts(self) -> None:
@@ -203,16 +229,36 @@ class Analysis:
               f"luminosity   : {self.luminosity:.2f} fb-1\n"
               )
 
+    # ===============================
+    # ========== PRIVATE ============
+    # ===============================
+    def __getbins(self, var_to_plot) -> Tuple[bool, Optional[tuple]]:
+        """
+        Returns special bins if variable input requires it. Returns None if not a special variable
+        :param var_to_plot: variable to choose bins from
+        :return: tuple (n_bins, start, stop) of bins or None
+        """
+        is_logbins = not any(map(var_to_plot.__contains__, self._not_log))
+        if not is_logbins:
+            # set bins for special variables
+            sp_var = [sp_var for sp_var in self._not_log if sp_var in var_to_plot]
+            if len(sp_var) != 1:
+                raise Exception(f"Expected one matching variable for spcial binning. Got {sp_var}")
+            return is_logbins, self._special_binning[sp_var[0]]
+        else:
+            return is_logbins, None
+
 
 if __name__ == '__main__':
     my_analysis = Analysis(root_path='../../data/mc16d_wmintaunu/*',
-                           cutfile='../../options/cutfile_taus',
+                           cutfile='../../options/wpt_cuts.txt',
                            lepton='tau',
                            force_rebuild=False
                            )
 
     # pipeline
     my_analysis.plot_with_cuts(scaling='xs')
-    my_analysis.gen_cutflow_hist(ratio=True, cummulative=True)
-    my_analysis.cutflow_printout()
+    my_analysis.make_all_cutgroup_2dplots()
+    # my_analysis.gen_cutflow_hist(ratio=True, cummulative=True)
+    # my_analysis.cutflow_printout()
     my_analysis.kinematics_printouts()
